### 2020-10-28 #2792 @whedon
[PRE REVIEW]: Cellocity: A Python package for analysis of confluent cell layer dynamics
**Submitting author:** @oftatkofta (<a href="http://orcid.org/0000-0002-8945-2665">Jens Eriksson</a>)
**Repository:** <a href="https://github.com/Oftatkofta/cellocity" target ="_blank">https://github.com/Oftatkofta/cellocity</a>
**Version:** 1.0.0
