### 2020-10-31 #2792 @jni
I have fairly low bandwidth this month but this is definitely in my wheelhouse! We just yesterday finished adding some PIV datasets to skimage! :joy: @Oftatkofta thanks for this work, apologies if I move a bit slow but I'll try to get reviewers lined up ASAP and get the process started!
