### 2020-11-04 #2818 @whedon
[REVIEW]: Cellocity: A Python package for analysis of confluent cell layer dynamics
**Submitting author:** @oftatkofta (<a href="http://orcid.org/0000-0002-8945-2665">Jens Eriksson</a>)
**Repository:** <a href="https://github.com/Oftatkofta/cellocity" target ="_blank">https://github.com/Oftatkofta/cellocity</a>
**Version:** v1.0.1
