### 2020-11-12 #2818 @jni
That seemed a bit blunt! :sweat_smile: @aplyer @rfezzani JOSS is testing a new "reminder" feature on reviews, please don't take it as too aggressive — it is just a kind ping to ensure things don't stall for months. :pray:
