### 2020-11-26 #2818 @jni
@rfezzani @aplyer It's been a very busy month for me and I've neglected my editorial duties. 😬 Probably a busy month for you also but I just wanted to make a small ping to surface this again to your attention. 😊
