### 2020-12-07 #2818 @rfezzani
Hi @jni, sorry again for my late answer.

`Cellocity` submission meets all JOSS requirements in terms of license, functionalities, documentation, and validation. I recommend  acceptance.
