### 2020-12-13 #2818 @jni
Thank you so much @rfezzani! @aplyer, apologies if you are busy — it is a busy time! — could you please let us know if you agree with @rfezzani's assessment? I see that most but not all of your checklist is ticked, so thank you for all the work you've done so far, and I look forward to wrapping this up! 😊
