### 2021-01-28 #2818 @jni
> MISSING DOIs
>
> - 10.1007/3-540-45103-x_50 may be a valid DOI for title: Two-Frame Motion Estimation Based on Polynomial Expansion

@Oftatkofta could you please add that DOI to that reference? Whedon appears to be accurate here. =)

Once that's added, could you please also make a new tagged release and Zenodo archive, and report back here the version number and the Zenodo DOI?
