### 2021-01-28 9ae43d2
Review changes
Added DOI for reference and correctet typo in paper
