+++ paper.bib
@@ -151,6 +151,7 @@
   year 			= {2003},
   series 		= {{LNCS} 2749},
   address 		= {Gothenburg, Sweden},
+  doi			= {10.1007/3-540-45103-x_50},
   month 		= {June-July}
 }
 
