### 2021-01-28 #2818 @Oftatkofta
@jni The new release has version 1.0.1, is submitted to Zendo and given the DOI 10.5281/zenodo.4475442 whedon seems satisfied with the DOI added to the reference.

On a personal note I would like thank you, the reviewers, and JOSS for a modern and smooth submission & review process.
