+++ paper.md
@@ -26,12 +26,12 @@ bibliography: paper.bib
 
 # Summary
 
-Studying the coordinated cell and tissue movements within confluent cell layers can give insights into diverse biological phenomena, such as wound healing, cancer cell metastasis, and host cell responses to infection. Cellular dynamics studies typically rely on analysis of time lapse images, generated by live cell microscopy. As a result of recent progress in stem cell- and developmental biology, it is now often possible to generate confluent cell layers of primary cells that recapitulate central features of intact tissues [@Dutton2019]. However, one disadvantage of such primary cell model systems is that it remains technically challenging to generate and image fluorescently labeled live samples. This complicates the use of segmentation and single-particle tracking as a means to study cellular movements [@Danuser2006]. Consequently, Optical Flow-based analysis of Bright field (black and white) microscopy images provides a competitive method to extract cell dynamics information from such sensitive and low-contrast experimental model systems [@Vig2016]. Because the study of confluent cell layer dynamics lies at the intersection of biophysics and cell biology, two research communities with different emphasis on mathematics, there is a need for a simple and easy-to-use framework that performs and visualizes Optical Flow analysis of microscopy data. We here present Cellocity, a Python package for Optical Flow and Particle Image Velocimetry (PIV) analysis, specifically catering to the bioimaging community.          
+Studying the coordinated cell and tissue movements within confluent cell layers can give insights into diverse biological phenomena, such as wound healing, cancer cell metastasis, and host cell responses to infection. Cellular dynamics studies typically rely on analysis of time lapse images, generated by live cell microscopy. As a result of recent progress in stem cell and developmental biology, it is now often possible to generate confluent cell layers of primary cells that recapitulate central features of intact tissues [@Dutton2019]. However, one disadvantage of such primary cell model systems is that it remains technically challenging to generate and image fluorescently labeled live samples. This complicates the use of segmentation and single-particle tracking as a means to study cellular movements [@Danuser2006]. Consequently, optical flow-based analysis of bright field (black and white) microscopy images provides a competitive method to extract cell dynamics information from such sensitive and low-contrast experimental model systems [@Vig2016]. Because the study of confluent cell layer dynamics lies at the intersection of biophysics and cell biology, two research communities with different emphasis on mathematics, there is a need for a simple and easy-to-use framework that performs and visualizes Optical Flow analysis of microscopy data. We here present Cellocity, a Python package for Optical Flow and Particle Image Velocimetry (PIV) analysis, specifically catering to the bioimaging community.          
 
 
 # Statement of need
 
-To date, Optical Flow and PIV analysis of microscopy data has primarily relied on different plugins for ImageJ, such as [PIV analyser](https://imagej.net/PIV_analyser), or on MATLAB scripts [@Vig2016]. However, these methods are limited in their accessibility, analysis capacity, metadata handling, and data visualization capabilities. Optical Flow and PIV are commonly used in the fields of fluid dynamics [@Taylor2010] and computer vision [@Bradski2000], and several open source frameworks exist to service these communities, e.g. [openPIV](http://www.openpiv.net/), [JPIV](https://eguvep.github.io/jpiv/index.html) and [OpenCV](https://opencv.org/) [@Bradski2000]. A corresponding framework for bioimaging and cell biology applications has so far been lacking.
[…]
