### 2021-01-28 #2818 @danielskatz
I've suggested some minor changes in https://github.com/Oftatkofta/cellocity/pull/21 - please merge this or let me know what you disagree with
